+++ .gitignore
@@ -22,3 +22,6 @@ database.db
 # Ignore logs and temporary files
 *.log
 *.tmp
+
+# Ignore sqlite instance
+instance/
