+++ app.py
@@ -1,101 +1,4 @@
-import os
-from datetime import datetime
-from flask import Flask, render_template, url_for, flash, redirect
-from flask_sqlalchemy import SQLAlchemy
-from dotenv import load_dotenv
-from forms import RegistrationForm, LoginForm
-from flask_wtf import FlaskForm
-
-load_dotenv()
-app = Flask(__name__)
-app.secret_key = os.environ.get('SECRET_KEY')
-app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
-db = SQLAlchemy(app)
-
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    image_profile = db.Column(db.String(20), nullable=False, default='default.jpg')
-    password = db.Column(db.String(60), nullable=False)
-    appointments = db.relationship('Appointment', backref='client', lazy=True)
-
-    def __repr__(self):
-        return f"User('{self.username}', '{self.email}', '{self.image_profile}')"
-
-class Therapist(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    therapist_name = db.Column(db.String(20), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    image_profile = db.Column(db.String(20), nullable=False, default='default.jpg')
-    password = db.Column(db.String(60), nullable=False)
-    specialization = db.Column(db.String(100), nullable=False)
-    appointments = db.relationship('Appointment', backref='doctor', lazy=True)
-
-    def __repr__(self):
-        return f"Therapist('{self.therapist_name}', '{self.email}', '{self.image_profile}', '{self.appointments}')"
-
-
-class Appointment(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    appointment_name = db.Column(db.Text, nullable=False)
-    appointment_date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    therapist_id = db.Column(db.Integer, db.ForeignKey('therapist.id'), nullable=False)
-    status = db.Column(db.String(20), nullable=False)
-
-    def __repr__(self):
-        return f"Appointment('{self.appointment_name}', '{self.appointment_date}')"
-
-
-users = [
-    {
-        'first_name': "Tony",
-        'last_name': "Mputhia",
-        'date_created': "26/02/2024",
-        'appointment': "12/04/2024",
-        'age': 28
-    },
-    {
-        'first_name': "Millyanne",
-        'last_name': "Adhiambo",
-        'date_created': "28/02/2024",
-        'appointment': "17/04/2024",
-        'age': 25
-    }
-]
-
-
-@app.route("/")
-@app.route("/home")
-def home():
-    return render_template("home.html", users=users)
-
-
-@app.route("/about")
-def about():
-    return render_template("about.html", title='About')
-
-@app.route("/register", methods=['GET', 'POST'])
-def register():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        flash(f'Account created for {form.username.data}!', 'sucess')
-        return redirect(url_for('home'))
-    return render_template("register.html", title='Register', form=form)
-
-
-@app.route("/login", methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        if form.email.data == 'tony@blog.com' and form.password.data == 'password':
-            flash(f'You have successfully logged in!', 'success')
-            return redirect(url_for('home'))
-        else:
-            flash(f'Login Unsuccessful! Check username and Password', 'danger')
-    return render_template("login.html", title='Login', form=form)
+from . import app
 
 
 if __name__ == "__main__":
